+++ core/server/server_scion.go
@@ -124,7 +124,7 @@ func runSCIONServer(ctx context.Context, log *slog.Logger, mtrcs *scionServerMet
 			continue
 		}
 		if flags != 0 {
-			log.LogAttrs(ctx, slog.LevelError, "failed to read packet", slog.Int64("flags", int64(flags)))
+			log.LogAttrs(ctx, slog.LevelError, "failed to read packet", slog.Int("flags", flags))
 			continue
 		}
 		oob = oob[:oobn]
